solution: small fixes from linux branch

+++ src/figure/formation_herd.cpp
@@ -15,7 +15,7 @@
 #include "sound/effect.h"
 
 static int get_free_tile(int x, int y, int allow_negative_desirability, int *x_tile, int *y_tile) {
-    unsigned short disallowed_terrain = ~(TERRAIN_ACCESS_RAMP | TERRAIN_MEADOW);
+    unsigned int disallowed_terrain = ~(TERRAIN_ACCESS_RAMP | TERRAIN_MEADOW);
     int tile_found = 0;
     int x_found = 0, y_found = 0;
